scripts/requirements.txt
numpy
pydantic
PyYAML
torch
tqdm
astropy
matplotlib
rich
scienceplots
